Reuse the same Vec for newly_acked
Currently every time an ack arrives quiche allocates a new Vec
to hold the newly acked packet descriptors. This is rather
expensive, and can be avoided by reusing the same Vec.
Sadly due to the way Recovery interface is implemented it can't
be integrated in Recovery without significant rewrite, and has
to be provided externally. Still the performance benefits are
pretty significant.