h3: convert control streams to non-incremental
The original idea was to set these streams as incremental so that
each one in the same urgency bucket had a fair share of bandwidth.
However, in reality there is little-to-no data being sent on these
streams. All the current behaviour does is create needless work in
the stream prioritizer for the entire length of a connection.

Changing to non-incremental reduces work and there should be no
noticable impact.