lib: progress through StreamIter from front to back
Previously, StreamIter next() would pop() the last stream ID and return
it, leading to a back-to-front progression.

Since we're stuck with SmallVec for internal data for performance reasons
we can't pop_front(). So instead, this change skips mutating the
StreamIter's internal data and instead uses a simple tracking index that
lets us move front-to-back.

This change helps to avoid cases where we had a correctly ordered
SmallVec of streams, and would need to reverse it before constructing a
StreamIter, just so that the iteration would make logical progressions.
It also addresses cases where we missed doing any reversal preparation.