automatically limit number of SCIDs that can be provded
By default we return the maximum possible number of SCIDs that can be
provided to the peer. However this is dangerous as each application will
need to limit this anyway to avoid potential DoS, and in practice there
doesn't seem to be much point in having different numbers of CIDs
between a client and server, so this change limits the number to
whatever is lowest between the clocally configured value and the peer's.

This also removes the `max_active_source_cids()` method as an
application can still fetch the raw TP value from the stats if it needs
to.