simplify ConnectionId
Currently a ConnectionId can either be a slice or a Vec. Because of this
declarations might need to also declare an explicit lifetime, which make
using the struct slightly annoying.

Instead just use a Vec all the time, and to avoid wasting memory when
cloning (since the CIDs are read-only anyway), wrap the internal Vec in
an Arc.