stream: properly tidy up reset streams that are read
When a peer resets their send side, quiche clears received data and inserts a a zero-length buffer
at the final size offset to ensure that the next stream_recv() call will return an Error to
indicate the reset.

Previously, when a reset error was returned, the zero-length buffer was not cleared, making it appear
that the stream was readable when everything else said it wasn't. This would confuse quiche's stream
limit accounting and lead to it not sending a MAX_STREAMS frame when it should have. This could lead
the peer running out of stream credits unexpectedly.

This change ensures the data is cleared in the terminal read state, and that MAX_STREAMS are emitted
once a stream has been fully handled.