lib: add getter for qlog streamer
If a quiche connection is configured to log using qlog (via set_qlog()
or set_qlog_with_level()) internally it creates a qlog streamer object
and using that to write out events.

This change adds a getter for the streamer, which allows library users
to also write their own events.