h3: allow the frame module to be public under a new feature
The quiche::h3::frame module is by default private because low-level
details are abstracted behind higher-layer APIs. However, for some
situations, such as testing, it can be helpful to have direct access
to frame types and their serialization/deserialization code.

This change adds a new 'internal' feature that when enables makes
the frame module public. There are no guarantees of API stability
for things exposed in this way.