Fixed incorrect lifetime annotation
The lifetime of the returned reference of the slice() and slice_last() methods was bounded by the lifetime of the method's receiver.
The methods in question return a subslice of a field stored inside the
receiver, whose lifetime may differ from that of the the method's receiver.
As such, the lifetime of the returned value should be bound to the lifetime of the receiver's field.