fixup! restore cwnd rollback only for in-flight packets
Previously we would only rollback cwnd updates due to spurious losses
for in-flight packets (while still updating threshold and stats for all
spurious losses) as that could mean there was no prior state to rollback
to.

The refactor made rollbacks happen for all spurious losses, so this
restores that (we still want to update threshold and stats for not
in-flight spurious losses too).