lib: fix resetting of streams with buffered but unsent data
An application can buffer stream data but not have the opportunity to prepare
the packets that would send that data. Within this window of time, the stream
might get reset either in reaction to receiving a STOP_SENDING frame, or due to
the application resetting the stream themselves via stream_shutdown(). Once this
occurs, no stream data will be sent. Instead, a RESET_STREAM will be emitted
when packets are next prepared that contains a final size of whatever bytes
_had_ been sent.

Quiche needs to keep a reference to such a stream for as long as it is "active".
For a bidi stream, that means that the receive side has been finished, and the
send side closure (RESET_STREAM) has been acknowledged and validated. Only when
both these conditions are true will the stream be collected. Stream collection
is used in stream credit accounting, which is used to for determining when to
emit a MAX_STREAMS frame.

Prior to this change, resetting a stream with buffered but unsent data would
lead to some internal checks failing when processing RESET_STREAM
acknowledgement. That led to not collecting the stream and not updating
MAX_STREAMS when it was expected.