generate packet numbers from a single sequence
Instead of keeping separate sequences for each packet number space, use
the same sequence.

This is needed to make it easier to support FIPS. This is because FIPS
mandates that the crypto module (e.g. BoringCrypto) has to validate that
the AEAD counter (which in QUIC corresponds to the packet number) is
stricly monotonically increasing (so that counters are not reused).

Because BoringCrypto saves the counter inside its own AEAD context, and
because new paths currently require starting from packet number 0 again,
the FIPS requirement would require us to maintain separate AEAD contexts
for each path, which we currently don't do (and would probably be messy
to implement).

This reverts commit 40e2433300d5bf57e3046a09e81974effb4697d7.