Add pointers about idle timeout in documentation
I spent an evening debugging an application, where both ends of the
connection were timing out at roughly the same time. I mistakenly
assumed that 'on_timeout' would cause ping packets to be sent to keep
the connection alive, but after reading through the RFC, it dawned on
me that applications are responsible for keeping the connection alive,
and Quiche merely detects the timeout. This design does make sense, but
a remark like the one I'm adding here would have saved me a lot of time,
so I hope it can be helpful to others in the future.