h3i: make Client a trait object
We originally had plans to completely deprecate the sync client once the
async one was finished, but we've since found a use for the sync client.

This gets rid of the ugly hacks we originally implemented to speed up
development by converting all sync/async conditional compilation into a
trait object. This means that we can simply implement the Client trait
for the async client