h3i: implement close trigger frames
Close trigger frames allow the user to specify a list of frames that the
h3i client expects to receive over a given connection.

If h3i sees all close triggers over the course of the connection, it
will pre-emptively close the connection with a CONNECTION_CLOSE frame.
If h3i does _not_ see all of the triggers, the resulting
ConnectionSummary will contain a list of the missing triggers for future
inspection.

This gives users a way to close tests out without waiting for the idle
timeout, or adding Wait/ConnectionClose actions to the end of each test.
This should vastly speed up test suites that have a large number of h3i
tests.