h3i: API cleanup
h3i's assertion API philosophy is to offload as much of the processing
as possible to the user, for maximum flexibility. As such, we shouldn't
expose a function that shows all HEADERS - rather, we should just allow
the user to provide their own lambdas.

If, in the future, we want to provide convenience methods for _all_
frame types, we can do so. However, we should think more on where we
want to draw that line.

This also changes StreamMap's backing store from a HashMap to a BTreeMap
to make inter-stream ordering deterministic when iterating over all
received frames.