crypto: replace ring HKDF API
Both BoringCrypto and OpenSSL provide APIs for HKDF, though not quite as
nice as ring's, but no point in pulling a whole dependency just for
that. This is also required (but not sufficient) for FIPS.

The change involves replacing the uses of `Salt` and `Prk` types exposed
by ring, with raw `Vec<u8>` to store intermediate secrets, which can
then be passed to BoringCrypto and OpenSSL via FFI.

Note that ring is still used to generate header protection keys, which
will be replaced next.