h3i: Improve ConnectionSummary stream_map serialization
The ConnectionSummary can be serialized to help other tooling consume
the output of h3i. Previously we didn't support the seralization of
DATA frame payloads. Furthermore, we were a bit inconsistent with how
we serialize frames with binary payloads.

This change adds support for logging DATA frame payloads and ensures
that all frames consistently have a `_len` field to indicate the true
length of the frame. Documentation has been added with a pointer to
the hexstring qlog definition that we use here.

Closes #1917