lib: Fix overflow when detecting reserved transport params
In cases where a QUIC connection's transport parameter has an ID of less
than 27, checking whether that transport parameter is reserved will
result in a panic caused by an attempt to store a negative value into a
variable with an unsigned type.

Handling transport parameters with ID of less than 27 specially will
prevent this from happening.

Signed-off-by: Will Hawkins <hawkinsw@obs.cr>