h3i: more workspace dependencies
This bumps url to match the versions specified in the workspace.

It also appears that the `serde_with` dependency was entirely
unnecessary, as the `serde_as` use in the code never actually worked
properly and it triggers a warning with newer versions
(https://github.com/jonasbb/serde_with/pull/502) so I just removed it.