tokio-quiche: expose quiche settings
50fb8dd97d2dc3ed4ee13be3326886a28c5b4dbe was reverted because we didn't
want to make a new major version, and the only settings that it exposed
were required for the async h3i client.

This commit re-enables the settings from the above commit, and adds the
rest of the relevant quiche settings to `QuicSettings`. The only field
which diverges from Quiche defaults is `enable_pacing`, which is `true`
by default in quiche but false in tokio-quiche. This seems fine since
tokio-quiche also requires pacing-compatible `SocketCapabilities`.

Note that this commit also adds a `#[non_exhaustive]` to `QuicSettings`
so that future settings additions don't force us to create a new major
version - however, that means that we'll need a major version here