tokio-quiche: expose max bandwidth and max loss percentage as stats
Currently these values are only exposed as metrics, but it can be useful
to collect them on a per-connection basis (rather than aggreagte
metrics), so expose them via audit log stats.

Regarding max loss percentage, the underlying value is `f32` but it's
converted to integer (`u8`) to allow using atomic operations like for
the rest of the `QuicAuditStats` fields.