tokio-quiche: always set UDP_SEGMENT cmsg when calling sendmsg() (#2060)
Current logic avoids setting the UDP_SEGMENT cmsg on sendmsg() calls
when sending a single packet over the wire. However, there are niche
cases where it's desirable to update the segment size after a socket has
been created. If the updated segment size is greater than what the
socket was initialized with, and we omit the UDP_SEGMENT cmsg, then the
kernel will fall back to the segment size that is configured on the
socket and erroneously truncate the packet on the way out.

We now set the UDP_SEGMENT cmsg on each packet no matter what.

To help avoid this footgun, we now initialize GSO on the socket with the
maximum possible segment size.