tokio-quiche: round-robin send over all network paths
Currently tokio-quiche depends on the `send()` method to pick the
network path to send packets on. This can mean that only packets for a
single path end being flushed to the network.

In addition to that, the destination address for `sendmsg()` calls is
fixed to the peer's initial address, initialized when a connection is
established.

These issues end up preventing connections from migrating correctly when
the peer's address changes (e.g. due to NAT rebinding) because packets
for any new path would have to be flushed too, and in any case the
original peer address would be used to send packets to anyway, rather
than the new address the peer migrated to.

This change fixes both of those issues, but note that while this allows
"passive" migration of peers, it does not allow "active" migration,
where a peer explicitly decides to switch network path by picking a
new connection ID. Supporting that would require providing additional
connection IDs on our part too, which could be implemented in the
future, but is out of scope of this change.

The test added is somewhat complicated, mostly because it needs to use
"plain" quiche to implement the migrating client, such that we can
control which address packets are sent from, but it simply uses 2
sockets bound to different addresses to simulate the client migrating
(similarly to what `quiche-client` does).