quiche: add is_handshake_confirmed API
`is_established()` will return `true` as soon as the handshake is
considered complete, but that will miss [post-handshake auth]. This is a
potential footgun as it could lead to applications mistakenly assuming
that post-handshake authorizations like mTLS are complete, when they're
not.

`is_handshake_confirmed` is a more explicit signal that the
QUIC connection considers the full handshake to be complete.