tokio-quiche: exit Handshake stage when handshake is confirmed
This makes exiting the handshake stage more aligned with the design of
the AOQ trait.

Take, for example, an mTLS client/server. The server could send a
CertificateRequest, but then drop through and send a Finished in the
same flight ([1]) since mTLS is considered [post-handshake
authentication]. This would cause a tokio-quiche-based mTLS client to
exit the handshake as soon as it sees the Finished, before it ever sends
its certificate. This could lead to cases where the AOQ sends data into
the void despite a pending handshake failure.

By instead exiting the handshake stage when the handshake is confirmed,
we can ensure that the AOQ only begins running when the handshake is
truly complete. This more closely matches the intended design of the AOQ
trait, which should only ever start handling packets when the entire
handshake is complete.

For servers, this is essentially a no-op change since
flipping `handshake_confirmed` to true is gated on
`handshake_completed`, which is the same as `is_established()`. For
clients, we now consider the handshake complete upon receiving a
`HANDSHAKE_DONE` frame rather than a TLS `Finished` message.

[post-handshake authentication]: https://datatracker.ietf.org/doc/html/rfc8446#section-4.6.2