tq: add monitoring of time taken to flush response headers
Sending headers in quiche means buffering from the application
layer into quiche. It is not always possible to do this due to
capacity at any moment in time.

Tokio-quiche currently sends response headers by processing the
writable frame queue (process_write_frame()) and acting on an
OutboundFrame::Headers. This supports informational headers,
final headers, and trailing headers. Each type might fail,
which results in leaving the OutboundFrame on the queue to be
retried later.

This change adds some time tracking to each stream. If the first
attempt to send headers fails, the time is marked. On completion,
a duration is calculated and the stream's aggregate flushing
time is incremented. This ensures retrying of any/all response
messages is accounted for.