h3: clear streams when send finishes before recv
h3 streams are meant to be dopped from the internal stream map when both
the send and recv side of the stream are finished (this means both
sending and received a `fin` for example).

However currently this only works if the recv side is finished first,
and the clean-up happens on the next `send_body()` or `send_response()`
call. When the opposite happens (e.g. `fin` is sent before receiving it)
the the stream leaks.

This change adds logic to properly check for a stream being completed in
that case as well, which involves tracking a "local finished" state for
each stream.

When returning a `Finished` event, the corresponding stream is checked
again and if its send-side is also finished then the stream's state is
cleared.

Note that this change doesn't cover the case of stream being reset by
the local application itself, as the h3 layer doesn't get any signal
about that happening right now, so it will need follow-up changes to
be fixed.