tokio-quiche: ignore unrecognized cmsgs
Router currently processes a variety of cmsgs. It returns an EINVAL if
it encounters a cmsg it doesn't care about. This means that we won't
create an Incoming or send the packet to the corresponding IOW.

Rather than black-holing packets with cmsgs we don't care about, we
should just no-op.