Bump nix to 0.30.1, bump mio to 1
This is required to give us access to previously-private fields on
UnknownCmsg.

We stop `clear()`ing the cmsg buffer because nix 0.30.1 transforms it to
a `msghdr` based on the buffer's length, whereas previously nix would do so
based on the buffer's capacity. Additionally, it was only cleared on each
call to `poll_recv_and_rx_time`, rather than before each `rcvmsg` call,
so any fears of leftover data should be alleviated.