tokio-quiche: allow users to read SO_MARK control messages
This allows users to read `SO_MARK` control message values on incoming
UDP packets. By extension, it attempts to set `SO_RCVMARK` on the
listening socket.

Note that this will only store the first `SO_MARK` value received on a
given connection. We grab the value from the packet which kicks off the
Handshake IOW.