tokio-quiche: better RESET_STREAM and STOP_SENDING handling
TL;DR:
* When receivign RESET_STREAM or STOP_SENDING, respond with RESET_STREAM
  and STOP_SENDING, respectively.
* When InboundFrame or OutboundFrame channels are closed without a fin,
  send STOP_SENDING or RESET_STREAM to close that half of the stream.
* Add a testing framework similar to quiche::test_utils::Pipe to allow
  unit testing of H3Driver.

This contains two changes to tokio-quiche's handling of "abnormal" stream
termination via RESET_STREAM and STOP_SENDING. A stream is only fully
completed if its final size in both directions is know and the final size
is only communicated via a fin or RESET_STREAM frame. Note that quiche
will automatically reply with a RESET_STREAM when it receives a STOP_SENDING
frame.

That means that receiving a RESET_STREAM or STOP_SENDING only shuts down one
direction of the stream, so tokio-quiche must also close the other direction.
This PR does so by responding with RESET_STREAM or STOP_SENDING. Note, that
I chose RequestCancelled as the error code in these messages, since I think
that's a better fit than simply echoing back the error we receivd from the
peer (since it's our choice to shut down the stream, we could have also continued
normally until a fin).

When the H3Driver detects that either the InboundFrame or OutboundFrame channels
have been closed without a fin, the driver shuts down that direction. This is
necessary to allow quiche to release the streams resources (flow control credit).
Since the channel is gone, it's otherwise impossible to finish the stream.