tokio-quiche: store latest PRIORITY_UPDATE value on incoming request
Priority signals take the form of HTTP headers and/or frames. Some clients
only send frames, while some clients can both. At the point in time a server
reads request HEADERS, a client may sent a PRIORITY_UPDATE frame that
should take precedence.

With this change, we read take the last PRIORITY_UPDATE processed by quiche,
and attach it to the IncomingH3Headers so that Tokio-Quiche-based apps can
use it if they wish.