task-killswitch: run worker as a separate thread
task-killswitch's operation queue is backing up in cases where many
tasks are spawned and deleted in quick succession on a busy tokio
runtime. Because of tokio's cooperative scheduling, a single poll of the
worker task can only process up to 128 messages before being put back at
the back of the thread's work queue.

Provided sufficient operations, the worker task is entirely CPU-bound,
so simply giving it more cooperative budget comes at the cost of
starving latency-sensitive tasks. Instead, we spawn a separate thread to
run the worker on. Additionally, we switch to `recv_many()` calls to
amortize some of the cost of receiving operations from the queue.