tokio-quiche: boost the priority of server streams prior to writing response headers
The timely delivery of response headers is important for HTTP because the metadata
can inform client behaviour such as canceling a request (e.g., seeing the content
is longer that desired) or initating new fetches (e.g., by receiving link headers).

This change boosts the priority of a stream at the moment the request has been
read, so that stream the stream scheduler and iterators can prefer streams in this
state. Tokio-quiche writes headers using the `with_priority` variants which allow
each individual stream priority to be tuned as instructed by the application
using TQ.