tokio-quiche: make wait_for_quiche cancellation-safe
The `wait_for_quiche` future waits for the handshake to generate data.
Once the handshake has done so, the future will attempt to flush the
bytes to the network, yielding if the socket isn't writable. Since it's
called in the `work_loop` `select!` statement, it's subject to
cancellation if a Quiche timer fires or a packet is received.

The issue is that each iteration of the top-level `work_loop` will call
`gather_data_from_quiche_conn` at least once, and
`gather_data_from_quiche_conn` resets the bytes_written counter on each
call. If the handshake generates data, and a timer/packet resolves the
`select!` while the socket isn't writable, we'll drop the
`wait_for_quiche` future without flushing and unconditionally set
bytes_written back to 0. This means we'll either delay or never send the
generated handshake data over the network.

This commit splits the flush operation out of the select! arm to make
the `wait_for_quiche` future cancellation-safe. If a timer/packet
resolves the `select!`, we must not have written any bytes and are
therefore good to drop the future. We'll next attempt to generate data
when we receive a CRYPTO frame or call `send_on_path`.