tokio-quiche: Add SCID to retry token HMAC
Due to tokio-quiche's stateless address validation via RETRY packets, we
rely on the client to reflect our chosen SCID back to us. We should make
sure that we actually generated that SCID by binding the retry token to
the SCID.

I didn't want to duplicate the SCID inside the token, so I refactored
the code a bit to only include data that is not available from
elsewhere. This means we also don't need to include the client IP
address in it anymore.