tokio-quiche: Remove socket_cookie from API
We can support socket-specific CID generators by moving the trait object
from the top-level entrypoints to `QuicListener`, which exists
separately for each socket. This allows the implementor full flexibility
in terms of what properties to consider when generating CIDs for a given
socket.

To avoid a type parameter in `QuicListener`, we require the user to pass
us an `Arc<dyn ConnectionIdGenerator>` (which is the internal
representation we want to use anyway.) This also allows users to share
a generator across all of their sockets.