tokio-quiche: Refresh connection IDs in IoWorker::work_loop
We aim to always provide as many CIDs to the peer as the local+remote
transport parameters allow. At the same time, retired connection IDs are
unmapped to keep the ConnectionMap compact.

To produce new connection IDs, we wire up the socket's
`Arc<dyn ConnectionIdGenerator>` all the way to the IoWorker parameters.